Projectmanagement for QuiApp
This is a file for the MPM/GANTT Methode to track changes within the project tasks. To use this file u need the Freeware ProjectLibre. After the Download of the App via https://sourceforge.net/projects/projectlibre/ u can open the file in the ProjectLibre App and overview a GANTT/MPM to see how the project is timed and which tasks have been added for example.